Move node specific code into directory
